ref: move mixin and decorator to keybindings.py
